requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
lxml==4.9.3
undetected-chromedriver==3.5.4
schedule==1.2.0
APScheduler==3.10.4

